backend/requirements.txt
django==3.2.8
djangorestframework==3.12.4
django-cors-headers==3.10.0
psycopg2-binary==2.8.6
dj-database-url==0.5.0
python-dotenv==0.19.1